requirements.txt
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coverage==7.6.10
dnspython==2.7.0
Flask==3.1.0
Flask-Cors==5.0.0
# Editable install with no version control (flaskr==1.0.0)
#-e d:\vscode\flask_tutorial
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
pymongo==4.10.1
pytest==8.3.4
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.1.0
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.44.0
gunicorn==20.1.0
